Move service to services.py
