+++ homeassistant/components/shelly/services.yaml
@@ -1,0 +1,12 @@
+get_kvs:
+  fields:
+    device_id:
+      required: true
+      selector:
+        device:
+          integration: shelly
+    key:
+      required: true
+      example: "my_key"
+      selector:
+        text:
